notificaciones
notifica con la app abierta queda probarlo

+++ nexo_lab_app/lib/screens/chats/chat_detail_page.dart
@@ -376,7 +376,7 @@ class _ChatDetailPageState extends State<ChatDetailPage> {
       final target = (index * 88.0).clamp(
         0,
         _scrollController.position.maxScrollExtent,
-      );
+      ).toDouble();
       _scrollController.animateTo(
         target,
         duration: const Duration(milliseconds: 300),
